feat: move data fetching down to the components that need it, thus isolating which parts of your routes should be dynamic.
