+++ next.config.ts
@@ -1,7 +1,9 @@
 import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
-  /* config options here */
+    experimental: {
+        ppr: 'incremental'
+    }
 };
 
 export default nextConfig;
